requirements.txt
crewai==0.86.0
crewai-tools==0.17.0
ipywidgets
matplotlib
pandas
seaborn
tqdm
plotly
streamlit
poetry
tiktoken
azure-ai-documentintelligence
azure-search-documents
python-dotenv
azure-identity
PyMuPDF
faiss-cpu
autogen-core
autogen-ext
autogen-agentchat
autogen
yfinance
pytz
numpy
requests
bs4
docling